+++ src/View/Helper/HtmlHelper.php
@@ -96,10 +96,16 @@ class HtmlHelper extends CakeHtmlHelper
     {
         if (!empty($options['tooltip'])) {
             $options = optionValues(['data-toggle' => 'tooltip'], $options);
-            $options['title'] = $options['tooltip'];
+            $options['title'] = trim(h(strip_tags($options['tooltip'])));
+
+            if (!empty($options['tooltip-align'])) {
+                $options = optionValues([
+                    'data-placement' => $options['tooltip-align'],
+                ], $options);
+            }
         }
 
-        unset($options['tooltip']);
+        unset($options['tooltip'], $options['tooltip-align']);
 
         return $options;
     }