+++ tests/TestCase/Core/PluginTest.php
@@ -31,6 +31,18 @@ use MeTools\Core\Plugin;
 class PluginTest extends TestCase
 {
     /**
+     * Teardown any static object changes and restore them
+     * @return void
+     */
+    public function tearDown()
+    {
+        parent::tearDown();
+
+        Plugin::unload('TestPlugin');
+        Plugin::unload('AnotherTestPlugin');
+    }
+
+    /**
      * Tests for `all()` method
      * @return void
      * @test
@@ -85,11 +97,19 @@ class PluginTest extends TestCase
         $result = Plugin::path('MeTools', 'config' . DS . 'bootstrap.php');
         $this->assertEquals($expected, $result);
 
-        $result = Plugin::path('MeTools', 'config' . DS . 'bootstrap.php', true);
+        $result = Plugin::path(
+            'MeTools',
+            'config' . DS . 'bootstrap.php',
+            true
+        );
         $this->assertEquals($expected, $result);
 
         //No existing file
-        $result = Plugin::path('MeTools', 'config' . DS . 'no_existing.php', true);
+        $result = Plugin::path(
+            'MeTools',
+            'config' . DS . 'no_existing.php',
+            true
+        );
         $this->assertFalse($result);
 
         $result = Plugin::path('MeTools', [