+++ tests/TestCase/View/Helper/DropdownHelperTest.php
@@ -54,6 +54,7 @@ class DropdownHelperTest extends TestCase
     public function tearDown()
     {
         parent::tearDown();
+
         unset($this->Dropdown, $this->Html, $this->View);
     }