+++ tests/test_app/Plugin/TestPlugin/src/Utility/ParentTestClass.php
@@ -20,12 +20,15 @@
  * @license     http://www.gnu.org/licenses/agpl.txt AGPL License
  * @link        http://git.novatlantis.it Nova Atlantis Ltd
  */
+namespace TestPlugin\Utility;
 
-require_once 'functions/global.php';
-require_once 'functions/html.php';
-require_once 'detectors.php';
+class ParentTestClass
+{
+    public function firstParentTestMethod()
+    {
+    }
 
-//Sets the default MeTools name
-if (!defined('METOOLS')) {
-    define('METOOLS', 'MeTools');
+    public function secondParentTestMethod()
+    {
+    }
 }