+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # 2.x branch
 ## 2.10 branch
+### 2.10.1
+* fixed serious bug for `BBCodeHelper::remove()` method: now it doesn't remove
+    all the HTML code, but only the BBCode code;
+* added `getChildMethods()` global function;
+* added support for tooltips alignment;
+* fixed bug for tooltips with quotes or code;
+* removed "bold" style for shells.
+
 ### 2.10.0
 * `HtmlHelper` class has been improved. The `tip` option has been replaced
     with `help`. Checkboxes are displayed according to the browser;